requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
requests==2.32.5
pydantic==2.9.2
python-dateutil==2.9.0.post0
dateparser==1.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10

# LangChain stack
langchain==1.1.0
langchain-core==1.1.0
langchain-community==0.4.1
langchain-google-genai==3.2.0
langgraph==1.0.4